split deploys

+++ lib/capistrano/tasks/frontend.rake
@@ -1,12 +1,23 @@
 # frozen_string_literal: true
 
+namespace :deploy do
+  task :frontend do
+    ENV['deployments'] = 'frontend'
+    Rake::Task['deploy'].invoke
+  end
+end
+
 after 'deploy:updated', :build_frontend do
+  next unless ENV['deployments'].split(',').include? 'frontend'
+
   run_locally do
     execute 'cd client && NODE_ENV=production NPM_CONFIG_PRODUCTION=true npm run build'
   end
 end
 
 after 'deploy:updated', :copy_frontend do
+  next unless ENV['deployments'].split(',').include? 'frontend'
+
   run_locally do
     execute "sed  -i '' '/eileen-and-dans-wedding/{N;N;d;}' ~/.aws/credentials"
     execute 'cat .credentials.aws >> ~/.aws/credentials'
@@ -20,6 +31,8 @@ after 'deploy:updated', :copy_frontend do
 end
 
 after 'deploy:updated', :remove_frontend_build do
+  next unless ENV['deployments'].split(',').include? 'frontend'
+
   run_locally do
     execute 'cd client && rm -rf build/'
   end
